fix(zero-cache): improve handling of ownership changes in the change-streamer (#4994)
When an ownership change is detected from the CDC DB, the ChangeStreamer
generally exits immediately.

There is evidence of a race condition in which the change-streamer loop
does not exit and is instead blocked, causing the server to continue to
run and, in turn, wreak mild havoc resulting from multiple
replication-managers making litestream backups and purging records from
the change log:

User report:
*
https://discord.com/channels/830183651022471199/1425849436696875069/1426372487985234003

To protect against this:
* Improve the abort logic (which is one of the blocking calls within the
loop) to be non-blocking, leveraging the currently tracked `watermark`
instead of waiting for the `storer` to retrieve the state (as it may
have stopped).
* Prevent any change log purging from the `storer` if the ownership has
changed.