chore(planner): factor sql building out of table source (#5004)
The planner needs to build select statements for a connection in
addition to table source needing to do this.

The planner will prepare statements to send to
`[sqlite3_stmt_scanstatus](https://sqlite.org/c3ref/stmt_scanstatus.html)`
which will estimate the cost of the query.

Share the same code.