fix(zero-cache): avoid unique constraint violations during push caused by change history compression (#5030)
This addresses a previous possible constraint violation whereby changing
the columns of a unique key might violate the unique constraint of a
disjoint key in the same row, as the delete + set of the rows could have
happened in the wrong order.

User report:


https://discord.com/channels/830183651022471199/1422347526283071580/1422638434740207783

Test case added in https://github.com/rocicorp/mono/pull/5044 now
passes.

The approach taken here is to read all values which conflict with a
change log `set` operation, and to push deletes for them through IVM.
Assuming the change log is correct, all of these deleted conflict rows
would have been deleted/replaced by later change log `delete`/`set`
operations. The read of conflicting values is piggy backed on the
existing SQLite query for the prev value, but instead of just querying
for the one row matching the row key, it uses an OR filter to query for
rows matching any of the unique non-null keys.