chore(planner): ANDed exists are additive, not multiplicative (#5045)
Thought I addressed this while addressing `OR`. 

```ts
issue
  .whereExists('label')
  .whereExists('creator')
```

The costs should be added in this case. The same base rows (those from
issue) accumulate child rows in a pipeline rather than a cartesian
product forming.

I.e., we do not get a `creator` for every `issue,label` pairing.

---

# Implementation

`estimateCost` now returns a tuple:

```ts
{
  baseCardinality: number;
  runningCost: number;
}
```

When we hit an additive join, the cost computation is:

```ts
return {
  baseCardinality: parent.baseCardinality, // parent drives the pipeline
  // we process child rows for each row in the base cardinality
  runningCost: parent.runningCost + parent.baseCardinality * child.runningCost,
};
```

When we hit a nested join, the cost computation is:

```ts
return {
  baseCardinality: parent.baseCardinality,
  runningCost: parent.runningCost * child.runningCost,
}
```

## Detecting additive vs multiplicative

- If the parent of a join is another join, we're additive
- If the parent of a join is a source, we're multiplicative