chore(planner): consider `limit` when planning (#5048)
# Overview

The current limit algorithm is pretty simple:

1. Compute selectivity for each connection. This is
`est_rows_with_filters / est_rows_without_filters`
2. In each semi-join, divide the parent limit by the child selectivity
to get the "scan_est"

The intuition here is that if a child is _very selective_ we must
iterate more rows on the parent side before finding a match on the child
side.

**Example:**
`issue.whereExists('creator').limit(10)`

Every issue has a creator so the selectivity of the child is 1. This
means we have: `scan_est = 10/1 = 10`. I.e., we'll have to scan 10 rows
before fulfilling our limit of 10.

If it were:
`issue.whereExists('creator', q => q.name('ff'))`

The cost model will give us the estimated num of rows for a creator with
a given name. Maybe it is 10 out of 1,000. This means our selectivity
is: 10/1000 -> .01.

.01 selectivity means we have: `scan_est = 10/.01 = 1,000` -> we will
potentially have to scan 1,000 parent rows before finding 10 matches
from the child.

```ts
let scanEst = parentCost.baseCardinality;
if (this.#type === 'semi' && parentCost.limit !== undefined) {
  if (childCost.selectivity !== 0) {
    scanEst = Math.min(scanEst, parentCost.limit / childCost.selectivity);
  }
}
```

# Problem

The simple algorithm isn't quite complete. Think of this query:

```ts
user.whereExists('comments', q => q.whereCreated(gt(date)))
```

The selectivity on `comments` may be high but if users have created many
comments, this increases the likelihood of finding a match.

**Example**

Say we have users `1,2,3,4,5`

And a comment table. Below are user ids from the comment table, assuming
every user created 2 comments.

```
[4, 1, 3, 5, 2, 1, 5, 4, 3, 2]
```

If our filters decimated the table to only half the rows:

```
[1, 5, 4, 3, 2]
```

We have 0.5 selectivity so an estimated scan of `1/0.5 = 2` over `user`
to find a match.

But we see that all user ids are present in the set! We only scan 1 user
row to find a matching comment. This faster matching is because we have
a higher fanout from user->comment than 1.

So we need to do: `selectivity = 1 - Math.pow(1 - filterSelectivity,
fanout) = 0.75 -> scan_est = 1 / 0.75 = 1.33`
1.33 being closer to 1.

More on this is documented in
packages/zql/src/planner/SELECTIVITY_PLAN.md

**Future notes to self:**
- [ ] should we throw out filters that are not indexed by sqlite when
computing selectivity?
- [ ] we currently penalize the plan if it creates a temp b-tree. This
penalty is not applied in our `parentCost.limit / childCost.selectivity`
algorithm. Postgres has a separate cost which is "startup_cost" that
captures things like creating temp indices.
- [x] consider how join keys impact selectivity. Likely need to
recompute selectivity after constraint prop.