chore(planner): cost revamp: sqlite cost model: [1/n] (#5097)
Cost from SQLite is two components:

1. estimated rows
2. startup cost

Ideally all of our probes for `exists` checks would not have a startup
cost. Exists requires that its input is in sorted order, however. This
order must also be unambiguous so we append the primary key to it.

If the table is not indexed by both the constraint and primary key,
SQLite must scan all rows that match the constraint (rather than just
finding a single one) then sort them.

Startup cost captures this.

Also renamed "basedCardinality" to "rows" to be less confusing.