feat(zero): defineQuery (#4958)
# Introduce defineQuery as the unified replacement for syncedQuery and
syncedQueryWithArgs.

BREAKING CHANGE!

apps/zbugs updated as an example; internals related to query
identification and materialization were refactored.

## Conceptual Changes

There are more kinds of `Queries` now.

## Query

The main interface. It has builder methods like where, limit, and
related. This is the only one we expect people to interact with.

BREAKING CHANGE! Query no longer has `run`, `preload`, or `materialize`.
You need to call these methods on the zero instance or the transaction
object.

### RootNamedQuery

This is the concrete class you get when the result of `defineQuery` is
called.

```ts
const issues = defineQuery('issue', {validator: ...}, ({ctx, args}) => ...);
issues({id: 42}); // Returns a Query backed by a RootNamedQuery
```

RootNamedQuery does not have an AST. The AST is only available after the
context has been passed in. The function body passed into defineQuery is
not called until the context is passed in. The args, on the other hand,
are validated right away.

### ChainedQuery

Calling the chaining methods on the `RootNamedQuery` instance creates
chained queries. They are called “chained” because they keep a reference
to the parent query.

Chained queries do not have an AST either. Once you pass the context in,
we call the parent query and obtain the AST.