feat(zero): added cleanup for server errors and wiring (#5100)
Ready to review.

Features:
- Added `ApplicationError<T>` so API server code can surface structured
JSON metadata to clients. This wraps all thrown client errors and **can
also be used by the client themselves** to pass along structured error
output. This is also used to throw `app` errors when awaiting the
`.server` promise.
- Added a `QueryParseError` to `zql` synced queries which is thrown when
parsing args fails. **This is plumbed back to the callsite for a
query.**
- Deprecated `refetch` and added `retry` to `useQuery`

Fixes:
- Hardened `handleMutationRequest` and `handleGetQueriesRequest` to
parse bodies/params defensively, wrap unexpected failures, emit
`PushFailed`/`TransformFailed` responses with details.
- Added `DatabaseTransactionError` to differentiate between "unable to
open DB TX", which is not app-level, with an app-level error.
- Lots of tests.