feat(zero)!: defineQuery again  (#5103)
BREAKING CHANGE!

- This adds another type parameter to a lot of the types.
- This removes run, preload and materialize from `Query`.

Revert the revert.

The main issue was that builder.tableName returns a QueryImpl and not a
ZPGQuery. Then we tried to `.run()` that which fails.

Instead, we only use QueryImpl and let the DBTransaction handle running
the query.

Remove ZPGQuery and RunnableQuery.

This simplifies a lot of things :-)