feat: continuous change-log cleanup for single-node (#5135)
Adds a `ReplicaMonitor`, which is the single-node analog to the
litestream
[`BackupMonitor`](https://github.com/rocicorp/mono/pull/4263), with the
ReplicaMonitor checking the version of the (single) replica file instead
of monitoring the litestream backup.

The same `scheduleCleanup()` method is called on the change-streamer
from either monitor implementation, allowing the change streamer to
remain agnostic as to whether it is running in a single or multi-node
configuration.

User report:
*
https://discord.com/channels/830183651022471199/1433038658881196153/1433829126871191794