Make schema path optional for zero-cache-dev (#5137)
Prior to this PR, `zero-cache-dev` was requiring a schema path and if
one was not set, it defaulted to `schema.ts`. Because of this, warnings
were thrown about not being able to find the file or the permissions.

This PR checks to see if a schema path was given, and if so, does what
it normally would and deploy permissions. If a path was not given, it
skips all permissions deployment and just starts the zero cache.

This PR also removes the empty permissions config from the zbugs schema
which was in place because of this bug.

I have checked this all works against zbugs.

---------

Co-authored-by: Matt Kinnersley <kinnersleymatt@gmail.com>
Co-authored-by: Aaron Boodman <aaron@aaronboodman.com>