fix(replicache): fixed closing expo-sqlite DB before deleting (#5149)
We're seeing a crash with zslack when it's deployed via TestFlight. It
occurs regularly at 06:17am on my phone (I love waking up to app
crashes!).

I did some spelunking and think it's related to Replicache clearing
storage every 12 hours. It happens regularly at the same minute of the
day.


[crashlog.crash.log](https://github.com/user-attachments/files/23404883/crashlog.crash.log)

[crashlog.crash-2.log](https://github.com/user-attachments/files/23404890/crashlog.crash-2.log)

My hunch is that `expo-sqlite` is throwing a DeleteDatabaseException
from here:

https://github.com/expo/expo/blob/d0ce3b5621aed60f4915d4be6a22881e363779e9/packages/expo-sqlite/ios/SQLiteModule.swift#L510-L518

Since the delegate is just opened and then immediately deleted, this
throws an exception.

I also added some error black-holing around this deletion, so it doesn't
crash apps if it fails for other reasons on mobile devices.