chore(planner): verify more query forms. Add more relaxed testing (#5158)
The original tests aimed for 0.7 correlation between plan cost estimates
and actual execution costs.

We can't always get that due to missing stats.

SQLite assumes very low selectivity when stats are missing. This is an
ok default in that it biases towards running the user's query as they
wrote it rather than applying flips.

This update:
1. adds more tests
2. adds two new validation modes

One mode checks how close we are to the optimal plan. We're almost
always within 2x of the optimal plan.

Another mode checks if we're same or better than the no-flip form of the
query. This is always true except for 1 query where we are worse.

---

All tests have three validations:

```
['correlation', 0.8],
['within-optimal', 1],
['within-baseline', 0.08],
```

- the higher **correlation**, the better. Means plan estimates correlate
to observed costs
- the closer to 1 for **within-optimal** the better. Checks if chosen
plan is as fast as optimal plan. 2 means 2x slower.
- **within-baseline** - the lower the better. 0.5 means 2x faster than
baseline (no flip) query.