feat: support permanent volumes for replica storage (#5179)
Previously, a multi-node view-syncer relied on its replica file to _not_
exist at startup, always restoring it from litestream.

This was necessary in order to correctly handle situations in which the
replica file becomes invalid, e.g.:
* if the replication-manager resyncs to a new `replicaVersion`
* if the local replica is too old to be caught up by the
replication-manager

In such cases, the view-syncer would handle the error from the
replication-manager (i.e. change-streamer) by shutting down, relying on
the fact that the next instance would restore a fresher backup from
litestream.

> In particular, deleting the replica file before shutdown was not
straightforward because syncer processes still have open file handles /
read locks on the stale replica.

With this change, a new view-syncer instance can be started on the same
volume, possibly containing a stale replica file, made possible by:
* sending additional change log state in the response of the
view-syncer's initial `/snapshot` request
* using the change log state to determine if the (existing or restored)
local replica is compatible with the change-streamer; i.e. the same
conditions (based on the `replicaVersion` and `watermark`) that the
change-streamer checks on a new `/changes` subscription
* deleting the local replica and retrying the restore if it is stale

In particular, performing the check at this startup stage of the
view-syncer life cycle works because the syncer processes have not yet
been started and the replica file is still "owned" by a single process.

This allows the replica file to be hosted on (1) permanent volumes and
(2) semi-permanent volumes, such as those that last across container
restarts in kubernetes pods.