fix(zero-cache)!: fix desync caused by client/server inconsistency in keys used to identify rows (#5200)
This reverts commit 7b67d49670ce7ce73a63491ef5e9b0e682c15efd.

The Zero client keys rows by the primary key declared in its client
schema. The client schema can use any non-nullable unique index for the
primary key. This is validated by zero-cache and supported to allow
upstream tables without declared primary keys, and changing primary keys
in upstream.

Before this change the CVR keyed rows by the `union key`, which was the
union of all unique non-nullable indexes on the table. This was done so
that keys sent for row deletes included the columns for all valid
primary keys the client could be using. However, there is a problem with
this approach. The CVR and client being inconsistent on how to what key
is used to identify rows can lead to desync.

Consider a table
```
      CREATE TABLE uniques (
        id "TEXT|NOT_NULL",
        name "TEXT|NOT_NULL",
        _0_version TEXT NOT NULL
      );
      CREATE UNIQUE INDEX uniques_id ON uniques (id);
      CREATE UNIQUE INDEX uniques_name ON uniques (name);
```
The client uses `['id']` for it schema's primary key, and thus uses
`['id']` to key rows. The client's CVR on zero-cache uses `['id',
'name']` (the union-key, i.e. the union of all unique non-nullable
indexes) to key rows.

Lets say we have a row `{ id: 'i1', name: 'foo'}` and the row is updated
to `{ id: 'i1', name: 'bar'}`. The row `{ id: 'i1', name: 'foo'}` has
been synced to the client, the update happens while the client is not
connected. A catchup sync for this row will result in the CVR diffing
logic send patches to the client to add `{ id: 'i1', name: 'bar'}` and
to delete row with key `{ id: 'i1', name: 'foo'} as it is no longer
referenced. According to its union key, these rows have different
identities so it is valid to represent this situation as an add of the
new row, and a delete of the old row. On the client however, according
to its key of `['id']` these are all the same row, and so first `{ id:
'i1', name: 'foo'}` is replaced by `{ id: 'i1', name: 'bar'}` and then
delete row with key `{ id: 'i1', name: 'foo'} is interpreted as delete
row with `row.id == 'i1'` which incorrectly deletes row `{ id: 'i1',
name: 'bar'}`, leading to a de-sync.

To fix this we need to avoid the client and CVR keying rows with
different ids.

A CVR is tied to a client group which is tied to specific client schema.
As of protocol version 30 we include the client schema's primary key in
the `initConnection` message and associate it with the CVR. This change
updates the CVR, and server pipelines to key rows by the client schema's
primary key.

It also removes the KeyColumns logic added in
https://github.com/rocicorp/mono/pull/3582, as it is no longer possible
for the id used for the RowKey of a CVR to change. It is now constant
for a CVR.

Finally it removes /analyze-queryz endpoint. The endpoint would have
needed to be updated to include the client schema in its body, but
instead of updating it I've removed it as it is effectively replaced by
the inspector.

BREAKING CHANGE: Increases MIN_SERVER_SUPPORTED_SYNC_PROTOCOL to 30,
which corresponds to release 0.24, so clients need to be updated to at
least 0.24 before deploying a 0.25 zero-cache. Removes the
/analyze-queryz endpoint, which is replaced by the inspector utility.