fix(zql): defer completing orderBy(s) of ast until pipeline build and use client primary key (#5194)
Currently `QueryImpl` ensures all `orderBy` properties of a a zql ast
are 'complete', where 'complete' means include all primary key columns.
`MemorySource` and `TableSource` both assert that the sort of a
connection (which comes form the ast `orderBy` property properly contain
all primary key columns. As of
https://github.com/rocicorp/mono/commit/15d1b2a93f19d87719c4bee65f9e01961ccf6d0a
both `MemorySource` and `TableSource` use the client schema to determine
what the primary key columns are.

This can lead to a problem if the api-server is using a schema with a
different primary key for a table than a client. This is likely to arise
during a migration that is changing a tables primary key. Prior to this
change, and as of
https://github.com/rocicorp/mono/commit/15d1b2a93f19d87719c4bee65f9e01961ccf6d0a,
this could result in an assertion error, as the `TableSource` would
assert that the sort contains all of the clients primary key columns.
Before
https://github.com/rocicorp/mono/commit/15d1b2a93f19d87719c4bee65f9e01961ccf6d0a
`TableSource` would have accepted the sort if had all the columns of any
"potential primary key", i.e. any non-nullable unique index on the
table, but still the client and api-server disagreeing on the order
could result in logic errors during pagination.

This change defers completing the `orderBy` properties until an AST is
being built in `buildPipeline`. This way only the `orderBy` columns
explicitly added during query building are passed from api-server to
zero-cache, and then zero-cache completes them using the client's
primary key.