fix(zero-cache): pipeline-driver was in invalid state after ResetPipelinesSignal error (#5213)
Before this fix a pipeline-driver that was reset due to a
ResetPipelinesSignal would hit an assertion error on the next
advancement, causing the client to have to disconnect and reconnect.

Also adds a minimum to the time limit applied to advancements of 30ms,
so that we don't constantly reset pipelines for very fast hydrations.