fix(zero-cache): delay the replication-manager handoff for loadbalancer task registration (#5250)
Restore the original `replication-manager` behavior of delaying the
replication stream takeover to allow the task to be registered as a
healthy target by the load balancer (i.e. after a minimum number of
health checks). This fixes the temporary unreachability of the
replication-manager when the handoff happens before the load-balancer
has recognized the new replication-manager as healthy.

This original functionality was simplified away with the introduction of
auto-discovery (https://github.com/rocicorp/mono/pull/4335), since that
replaced the dns and proxying component, but never restored when
proxy-based routing was reintroduced in
https://github.com/rocicorp/mono/pull/4584 (and is now the recommended
configuration).

This new implementation is more compartmentalized than the original
implementation, encapsulating all of the logic in the
ChangeStreamerHttpService, so that the ChangeStreamerService itself is
agnostic to the details of health checks and startup delays.