chore(planner): stick values inline for planning (#5204)
https://www.sqlite.org/c3ref/stmt_scanstatus.html will not return
correct estimates when `?` are used. Even after binding, correct
estimates are not returned. Correct estimates are returned after calling
`step` on the query but this would require running the query during
planning. At least for 1 row of results.

A workaround is to put the value directly inline in the string during
planning `WHERE x = 'value'` which will return correct estimates.

This does that. And adds a bunch of tests to see if this improved or
made things worse. No observed change / returned estimates are better
but we're still picking the same plans.