chore(zero-cache): improve active user stats using profile IDs (#5324)
(Retake of #5322 with fixed migration code and accompanying test)

### `instances.profileID`

A new `profileID` is added to the CVR `instances` table to store the
stable `profileID` value from
https://github.com/rocicorp/mono/pull/5310, which addresses the
over-counting issue for apps that use memstore and create new client
group ids all the time.

The `profileID` of existing CVRs is backfilled to `cg${clientGroupID}`.

This value is updated similarly to how the `clientSchema` is updated;
namely, with initConnection the profileID is updated if different from
what is in the DB. This will allow a new `p{...}` profile ID from the
browser to replace the backfilled `cg{...}` id.

### `instances.deleted`

A new `deleted` column in the CVR `instances` row is used to mark a CVR
as deleted without deleting the row itself. The cvr-purger will delete
rows from all of the other tables (the bulk of the data), but leave the
`instances` row with the `deleted` bit set. It also resets other column
values like the `version` so that older zero-cache's will effectively
treat these loaded client groups as deleted.

CVR `instances` rows that are `deleted` are purged after 31 days instead
of using the shorter CVR purge interval (defaulting to 2 days).

### activity stats

* The existing `active_users_last_day` is left for continuity.
* The following stats count active users using distinct `profileID`
values from browsers (i.e. starting with "p"), thus avoiding the
overcounting issue:
  * `users_1da`
  * `users_7da`
  * `users_30da`
* Finally, the following hybrid metric counts active users using any
`profileID`, including the backfilled "cg" profile IDs. These will
eventually converge with the preceding "p"-based profile ID counts once
apps are upgraded to 0.25+
  * `users_1da_legacy`
  * `users_7da_legacy`
  * `users_30da_legacy`